tests: Fix "PHP Deprecated: Implicit conversion from float to int" warning
Happens on PHP 8.1+

> PHP Deprecated: Implicit conversion from float 6935.303211212158 to int
> loses precision in /tests/MinifyBenchmark.php on line 61

+++ tests/MinifyBenchmark.php
@@ -58,7 +58,7 @@ class MinifyBenchmark {
 		$this->outputStat( 'CSSMin::remap (example)', $total, $max, $iterations );
 	}
 
-	private function outputStat( string $name, int $total, int $max, int $iterations ): void {
+	private function outputStat( string $name, float $total, float $max, int $iterations ): void {
 		// in milliseconds
 		$mean = $total / $iterations;
 		$ratePerSecond = 1.0 / ( $mean / 1000.0 );
